sts/requirements.txt
Django==3.0.6
psycopg2==2.8.5
Pillow==7.1.2
Unidecode==1.1.1
Willow==1.3
beautifulsoup4==4.8.2
certifi==2020.4.5.1
chardet==3.0.4
django-modelcluster==5.0.1
django-taggit==1.3.0
django-treebeard==4.3.1
draftjs-exporter==2.1.7
html5lib==1.0.1
idna==2.9
l18n==2018.5
requests==2.23.0
soupsieve==2.0.1
wagtail==2.9
webencodings==0.5.1
xlsxwriter==1.2.8
gunicorn==20.0.4
aniso8601==7.0.0
graphene==2.1.8
graphene-django==2.10.1
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
rx==1.6.1
singledispatch==3.4.0.3
django-filter==2.2.0

